Rename 'add_list' to 'AddForms'
